Refactor file imports and add utility functions for AI review processing
