--- pr-review/src/read-from-env.ts
@@ -1,9 +1,0 @@
-export const getServiceKeyFromEnv = (): string => {
-  const { readFileSync } = require("fs")
-  const { resolve } = require("path")
-  const { parse } = require("dotenv")
-
-  const env = parse(readFileSync(resolve(process.cwd(), ".env")))
-
-  return env.AICORE_SERVICE_KEY || "missing service key"
-}
