+++ pr-review/test/index.spec.ts
@@ -1,6 +1,6 @@
-import { getServiceKeyFromEnv } from "src/helpers/read-from-env.ts"
 import { describe, expect, it } from "vitest"
 import type { Config } from "../src/domain/model/config.ts"
+import { getServiceKeyFromEnv } from "../src/domain/utils/read-from-env.ts"
 import { run } from "../src/main.ts"
 
 const getConfig = (): Config => {
